Revert "Moved examples and removed duplicated json-response (#350)"
This reverts commit cf707a7a9bc57882d43d47cf5b88a081e8148738.
