+++ edge-api-routes/cache-control/api/edge.ts
