+++ edge-api-routes/cache-control/tsconfig.json
@@ -1,0 +1,1 @@
+{}
