+++ edge-api-routes/json-response/.gitignore
@@ -1,0 +1,1 @@
+.vercel
