+++ edge-api-routes/json-response/README.md
@@ -1,0 +1,27 @@
+# JSON Reponse
+
+Build your API using Vercel Edge Functions (Experimental).
+
+## How to Use
+
+You can choose from one of the following two methods to use this repository:
+
+### One-Click Deploy
+
+Deploy the example using [Vercel](https://vercel.com?utm_source=github&utm_medium=readme&utm_campaign=vercel-examples):
+
+[![Deploy with Vercel](https://vercel.com/button)](https://vercel.com/new/git/external?repository-url=https://github.com/vercel/examples/tree/main/edge-api-routes/json-response&project-name=json-response&repository-name=json-response)
+
+### Running Locally
+
+Install the Vercel CLI:
+
+```bash
+npm i -g vercel
+```
+
+Then run the example at the root of the repository:
+
+```bash
+vercel dev
+```
