+++ edge-api-routes/json-response/api/edge.ts
