+++ edge-api-routes/json-response/public/index.html
