+++ edge-api-routes/json-response/tsconfig.json
@@ -1,0 +1,1 @@
+{}
