+++ edge-api-routes/query-parameters/tsconfig.json
@@ -1,0 +1,1 @@
+{}
