--- edge-functions/authed-proxy/tsconfig.json
@@ -1,1 +1,0 @@
-{}
