--- edge-functions/cache-control/tsconfig.json
@@ -1,1 +1,0 @@
-{}
