--- edge-functions/hello-world/tsconfig.json
@@ -1,1 +1,0 @@
-{}
