+++ edge-functions/json-response/.eslintrc.json
@@ -1,0 +1,4 @@
+{
+  "root": true,
+  "extends": "next/core-web-vitals"
+}
