+++ edge-functions/json-response/.npmrc
@@ -1,0 +1,2 @@
+# Enabled to avoid deps failing to use next@canary
+legacy-peer-deps=true
