+++ edge-functions/json-response/README.md
@@ -1,6 +1,21 @@
-# JSON Response
+---
+marketplace: false
+---
 
-Build your API using Vercel Edge Functions (Experimental).
+# JSON Response Example
+
+```ts
+import type { NextRequest } from 'next/server'
+
+export function middleware(req: NextRequest) {
+  return new Response(JSON.stringify({ message: 'hello world!' }), {
+    status: 200,
+    headers: {
+      'Content-Type': 'application/json',
+    },
+  })
+}
+```
 
 ## Demo
 
@@ -14,20 +29,28 @@ You can choose from one of the following two methods to use this repository:
 
 Deploy the example using [Vercel](https://vercel.com?utm_source=github&utm_medium=readme&utm_campaign=vercel-examples):
 
-[![Deploy with Vercel](https://vercel.com/button)](https://vercel.com/new/git/external?repository-url=https://github.com/vercel/examples/tree/main/edge-api-routes/json-response&project-name=json-response&repository-name=json-response)
+[![Deploy with Vercel](https://vercel.com/button)](https://vercel.com/new/git/external?repository-url=https://github.com/vercel/examples/tree/main/edge-functions/json-response&project-name=json-response&repository-name=json-response)
 
-### Running Locally
+### Clone and Deploy
 
-Install the Vercel CLI:
+Execute [`create-next-app`](https://github.com/vercel/next.js/tree/canary/packages/create-next-app) with [npm](https://docs.npmjs.com/cli/init) or [Yarn](https://yarnpkg.com/lang/en/docs/cli/create/) to bootstrap the example:
 
 ```bash
-npm i -g vercel
+npx create-next-app --example https://github.com/vercel/examples/tree/main/edge-functions/json-response json-response
+# or
+yarn create next-app --example https://github.com/vercel/examples/tree/main/edge-functions/json-response json-response
 ```
 
-Then run the example at the root of the repository:
+Next, run Next.js in development mode:
 
 ```bash
-vercel dev
+npm install
+npm run dev
+
+# or
+
+yarn
+yarn dev
 ```
 
 Deploy it to the cloud with [Vercel](https://vercel.com/new?utm_source=github&utm_medium=readme&utm_campaign=edge-middleware-eap) ([Documentation](https://nextjs.org/docs/deployment)).
