+++ edge-functions/json-response/package.json
@@ -1,7 +1,24 @@
 {
-  "private": true,
+  "name": "json-response",
+  "version": "1.0.0",
+  "repository": "https://github.com/vercel/examples.git",
+  "license": "MIT",
+  "scripts": {
+    "dev": "next dev",
+    "build": "next build",
+    "start": "next start",
+    "lint": "next lint"
+  },
+  "dependencies": {
+    "next": "canary",
+    "react": "latest",
+    "react-dom": "latest"
+  },
   "devDependencies": {
-    "@types/node": "^17.0.42",
-    "typescript": "^4.7.3"
+    "@types/node": "^17.0.14",
+    "@types/react": "latest",
+    "eslint": "^8.9.0",
+    "eslint-config-next": "canary",
+    "typescript": "4.5.5"
   }
 }
