+++ edge-functions/json-response/pages/api/json.ts
@@ -1,0 +1,5 @@
+import type { NextApiRequest, NextApiResponse } from 'next'
+
+export default function handler(_: NextApiRequest, res: NextApiResponse) {
+  res.status(200).json({ message: 'hello world!' })
+}
