--- edge-functions/query-parameters/tsconfig.json
@@ -1,1 +1,0 @@
-{}
